+++ ProyectoFinalRC/src/Components/Pages/AboutUs/AboutUs.jsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import AboutUsBanner from './AboutUsBanner/AboutUsBanner'
-import Card from "./Card"
+import Card from "./Cards/Card"
+import Card2 from "./Cards/Card2"
 import "./AboutUs.css"
 
 const AboutUs = () => {
@@ -12,8 +13,8 @@ const AboutUs = () => {
     <div className='Card'>
       <Card/>
     </div>
-    <div className='Card'>
-      <Card/>
+    <div className='Card2'>
+      <Card2/>
     </div>
     <div className='Card'>
       <Card/>
