--- ProyectoFinalRC/src/Components/Pages/AboutUs/Card.css
@@ -1,38 +1,0 @@
-.about-us-card {
-  display: flex;
-  align-items: center;
-  background-color: #f0f0f0;
-  border: 1px solid #ccc;
-  padding: 20px;
-  border-radius: 8px;
-  margin: 90px;
-}
-
-.profile-pic {
-  width: 120px; /* Ajusta el tamaño según sea necesario */
-  height: 120px; /* Ajusta el tamaño según sea necesario */
-  overflow: hidden;
-  border-radius: 50%; /* Para hacer la imagen redonda si es necesario */
-  margin-right: 20px;
-}
-
-.profile-pic img {
-  width: 100%;
-  height: auto;
-}
-
-.text-content {
-  flex: 1; /* Para que ocupe el espacio restante */
-}
-
-.about-us-card h2 {
-  color: #333;
-  font-size: 24px;
-  margin-bottom: 10px;
-}
-
-.about-us-card p {
-  color: #666;
-  font-size: 16px;
-  line-height: 1.6;
-}
