--- ProyectoFinalRC/src/Components/Pages/AboutUs/Card.jsx
@@ -1,27 +1,0 @@
-import React from 'react';
-import './Card.css'; // Estilos CSS para la tarjeta
-// Importa la imagen de perfil
-
-const Card = () => {
-  return (
-    <div className="about-us-card">
-      <div className="profile-pic">
-        <img src="" alt="Profile" />
-      </div>
-      <div className="text-content">
-        <h2></h2>
-        <p>
-          
-        </p>
-        <p>
-          
-        </p>
-        <p>
-          
-        </p>
-      </div>
-    </div>
-  );
-}
-
-export default Card;
